@@ -6,7 +6,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.17.3 <3.0.0"
+  sdk: "3.5.0"
 
 dependency_overrides:
   intl: ^0.20.1
@@ -61,6 +61,7 @@ dependencies:
   gql_http_link: ^1.1.0
   hive_flutter: ^1.1.0
   ferry_hive_store: ^0.6.0
+  logging: ^1.3.0
 
 dev_dependencies:
   flutter_test:
